requirements.txt
django-taggit==6.1.0
pyyaml==6.0.3
pymdown-extensions==10.21.2
nh3==0.3.4
markdown==3.10.2
psycopg-binary==3.3.3
psycopg==3.3.3
social-auth-app-django==5.7.0
django-bootstrap5==26.2
djangorestframework==3.17.1
